qrcode_decoder/requirements.txt
pandas
pyzbar
Pillow